Modified the 0-read_textfile.c file

+++ 0x15-file_io/0-read_textfile.c
@@ -16,7 +16,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	if (filename == NULL)
 		return (0);
 
-	buf = malloc(sizeof(char) * letters);
+	buf = malloc(sizeof(char) *(int) letters);
 
 	if (buf == NULL)
 		return (0);
